% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{‘accident_2015’}
\alias{‘accident_2015’}
\alias{accident_2015}
\title{Data on fatality}
\format{
A data frame with 30202 rows and 50 variables:
\describe{
  \item{STATE}{state number, as integers}
  \item{YEAR}{year of the fatalities}
  \item{LATITUDE}{year of the fatalities}
  \item{LONGITUDE}{year of the fatalities}
  ...
}
}
\source{
\url{https://www.coursera.org/learn/r-packages}
}
\usage{
accident_2015
}
\description{
A compressed dataset containing the fatalities per state and date in 2013
}
\keyword{datasets}
