% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read is a function that reads a csv table.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file that needs to be read.}
}
\value{
it returns the datafile that has been read transformed into a tibble
}
\description{
This is an internal simple function that reads a file when it exists and it converts it into a tibble
}
\note{
if the function is provided a filename that is not available it returns an error. One needs to be sure that one is in the right directory or that the provided filenames exist.
}
\examples{
fars_read(\code{filename})

}
