% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read external csv data}
\usage{
sample_read(path)
}
\arguments{
\item{path}{to the file}
}
\value{
a \code{tibble}
}
\description{
This is an internal simple function that reads a file when it exists and it converts it into a tibble
}
\examples{
path =  system.file("extdata","accident_2013.csv.bz2",package="testFars")
sample_read(path)
}
